switch to BaseSimpleActivity. add percentage and amount. change float to double for GMGraphViewModel. add PercentageUtil for GMTransactionGraphViewHelper. add tools context for fragment gm statistic transaction.

+++ tkpdseller/src/main/java/com/tokopedia/seller/goldmerchant/statistic/view/helper/model/GMGraphViewModel.java
@@ -11,7 +11,7 @@ public class GMGraphViewModel {
     public List<Integer> dates;
     public String title;
     public String subtitle;
-    public float percentage;
+    public double percentage;
     public int amount;
     public GMDateRangeDateViewModel dateRangeModel;
 }
